Исправление работы пагинации

+++ category.php
@@ -32,7 +32,7 @@ if ($lotsCount) {
 dbClose($dbConnection);
 
 $pages = [];
-$pagesCount = ceil($lotsCount / $lotsPerPage);
+$pagesCount = (int)ceil($lotsCount / $lotsPerPage);
 if ($pagesCount > 1) {
     $pages = range(1, $pagesCount);
 }
